[Cart] extract summary tests
